test(search): move command tests to module file
